fix: replace default input values on first keypress (#23)
## Motivation

When users reached cluster name or worker nodes input steps in the
create wizard, typing would append to the generated default value
instead of replacing it. This created confusing UX where users had to
manually delete the default before entering their desired value.

## Implementation information

Added `nameModified` and `workersModified` boolean flags to
`createFormModel` to track whether the user has started typing in each
field. On first keypress, the default value is replaced and the flag is
set. Subsequent keypresses append normally. When navigating back with
Esc, the modification flags are reset to allow replacing defaults again
if the user returns to that step.

## Supporting documentation

Resolves user-reported UX issue.

---------

Signed-off-by: Marcin Skalski <skalskimarcin33@gmail.com>